remove CenteredFlex.tsx and move ViewPlaceholder.tsx in layout.
