--- src/components/common/fragments/CenteredFlex.tsx
@@ -1,13 +1,0 @@
-import React, { FunctionComponent } from "react";
-
-interface CenteredFlexProps {
-    className?: string
-}
-
-export const CenteredFlex: FunctionComponent<CenteredFlexProps> = (props) => {
-    const {children, className} = props;
-
-    return (
-        <div className={className ? `centeredFlex ${className}` : "centeredFlex"}>{children}</div>
-    );
-};
