+++ src/components/layout/ViewPlaceholder.css
@@ -1,0 +1,8 @@
+.viewPlaceholder {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    height: 100%;
+    width: 100%;
+    color: var(--text-color);
+}
