+++ src/css/Davit.css
@@ -608,15 +608,6 @@
     transition: background-color var(--animation-time);
 }
 
-.viewPlaceholder {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    height: 100%;
-    width: 100%;
-    color: var(--text-color);
-}
-
 .inputLabel {
     display: flex;
     justify-content: center;
