+++ src/style/layout.css
@@ -39,3 +39,8 @@
 .flex-shrink {
     flex-shrink: 1;
 }
+
+.flex-center {
+    justify-content: center;
+    align-items: center;
+}
